refactor: move the size logic from the plugin wrapper
This is because eventually the PluginWrapper component will be replaced
by the generic one from app-platform.
If possible all custom logic should not be put in other components at
the app level.
